requirements.txt
# Core Agent Framework
langchain>=0.3.0
langgraph>=0.2.0
langchain-nvidia-ai-endpoints>=0.3.0
langchain-openai>=0.2.0

# Tools
e2b-code-interpreter>=1.0.0
tavily-python>=0.5.0

# Evaluation
datasets>=3.0.0

# Frontend
streamlit>=1.40.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0
